requirements.txt
python-telegram-bot==20.6
opencv-python==4.5.1
numpy==1.21.0